Restructured modules, simplified code
